custom image, typed methods and image integration into blog post
